Added Parent-Child Search using Elastic and Queueing

+++ services/elasticsearch.service.js
@@ -1,20 +1,20 @@
-import { client, elasticServiceConnection } from '../services/elasticServiceConnection.js';
-// import config from "../../config/local.json";
+import { client } from './elasticServiceConnection.js';
 
 const INDEX_NAME = "planindex";
 
 client.ping((error) => {
     if (error) {
-        console.trace('Elasticsearch cluster is down!');
-    } else {
-        console.log('Elastic search client is working fine!');
+        console.trace('Elasticsearch Cluster is down!');
     }
+    console.log('Elastic search client is working fine!');
 });
 
 let MapOfDocuments = {};
 let listOfKeys = [];
 
 const convertMapToDocumentIndex = async (jsonObject, parentId, objectName, parentObjId) => {
+    if (!jsonObject || typeof jsonObject !== 'object') return {}; 
+
     const valueMap = {};
     const map = {};
 
@@ -33,24 +33,23 @@ const convertMapToDocumentIndex = async (jsonObject, parentId, objectName, paren
     if (objectName === "plan") {
         valueMap["plan_join"] = {
             "parent": "",
-            "name": objectName,
+            "name": objectName
         };
     } else if (objectName.match(/^-?\d+$/)) {
         parentId = parentObjId;
         valueMap["plan_join"] = {
             "parent": parentObjId,
-            "name": "linkedPlanServices",
+            "name": "linkedPlanServices"
         };
     } else {
         valueMap["plan_join"] = {
             "name": objectName,
-            "parent": parentId,
+            "parent": parentId
         };
     }
 
     const id = `${parentId}:${jsonObject.objectId}`;
-    if (jsonObject?.objectId) MapOfDocuments[id] = valueMap;
-
+    if (!!jsonObject?.objectId) MapOfDocuments[id] = valueMap;
     return map;
 };
 
@@ -69,8 +68,8 @@ const convertToList = async (jsonArray, parentId, objectName, parentObjId) => {
 };
 
 const convertToKeysList = async (jsonArray) => {
-    const list = [];
-    for (const value of jsonArray) {
+    let list = [];
+    for (let value of jsonArray) {
         if (Array.isArray(value)) {
             value = await convertToKeysList(value);
         } else if (typeof value === 'object') {
@@ -82,6 +81,11 @@ const convertToKeysList = async (jsonArray) => {
 };
 
 const convertToKeys = async (jsonObject) => {
+    if (!jsonObject || typeof jsonObject !== 'object') {
+        console.error("Invalid or null jsonObject passed to convertToKeys:", jsonObject);
+        return {}; 
+    }
+
     const map = {};
     const valueMap = {};
 
@@ -97,7 +101,9 @@ const convertToKeys = async (jsonObject) => {
         }
     }
 
-    listOfKeys.push(jsonObject["objectId"]);
+    if (jsonObject["objectId"]) {
+        listOfKeys.push(jsonObject["objectId"]);
+    }
     return map;
 };
 
@@ -116,15 +122,20 @@ const postDocument = async (plan) => {
         }
         return { message: 'Document has been posted', status: 200 };
     } catch (e) {
-        console.error("Error", e);
+        console.log("Error", e);
         return { message: 'Document has not been posted', status: 500 };
     }
 };
 
 const deleteDocument = async (jsonObject) => {
+    if (!jsonObject || typeof jsonObject !== 'object') {
+        console.error("Invalid or null jsonObject provided to deleteDocument:", jsonObject);
+        return;
+    }
+
     listOfKeys = [];
     await convertToKeys(jsonObject);
-    console.log(listOfKeys);
+
     for (const key of listOfKeys) {
         client.delete({
             index: INDEX_NAME,
